requirements.txt
absl-py==0.15.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==4.2.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
blinker==1.8.2
brisque==0.0.16
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==3.30.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
commonmark==0.9.1
contourpy==1.1.1
cryptography==43.0.1
cycler==0.12.1
datasets==2.21.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
-e git+https://github.com/VinAIResearch/Anti-DreamBooth.git@410ad27f56289a6a57c1e9d68610ac1e586ec77f#egg=deepface&subdirectory=evaluations/deepface
diffusers==0.30.0
dill==0.3.6
dnnlib @ https://github.com/podgorskiy/dnnlib/releases/download/0.0.1/dnnlib-0.0.1-py3-none-any.whl#sha256=e789afb6365198754f3eedd7687d2c321f6a2daf8009eb94e46fb8b90217361d
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fairscale==0.4.13
filelock==3.13.1
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.2
flatbuffers==1.12
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==23.0.0
h5py==3.1.0
huggingface-hub==0.24.6
idna==3.7
ifaddr==0.2.0
image-quality==1.2.7
image-reward==1.5
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257397447/work
keras==2.13.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
libsvm==3.23.0.4
libsvm-official==3.35.0
lit==18.1.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mlp-mixer-pytorch==0.1.1
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.5
multiprocess==0.70.14
mypy-extensions==1.0.0
natsort @ file:///tmp/build/80754af9/natsort_1611692802956/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.0rc1
ninja==1.11.1.1
numpy==1.21.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==1.4.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
peft==0.12.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
piq==0.8.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
pyee==12.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pylibsrtp==0.10.0
Pympler==1.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyre-extensions==0.0.23
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytorch-triton==3.0.0+dedb7bdf33
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1705605076900/work
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
-e git+https://github.com/VinAIResearch/Anti-DreamBooth.git@410ad27f56289a6a57c1e9d68610ac1e586ec77f#egg=retina_face&subdirectory=evaluations/retinaface
rich==12.0.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semver==3.0.2
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.15.0
smmap==5.0.1
soupsieve==2.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.10.0
streamlit-webrtc==0.42.0
sympy==1.12.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.13
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.1
torch-dct==0.1.6
torch-utils==0.1.2
torchvision==0.19.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.66.5
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.44.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
validators==0.34.0
wandb==0.17.8
watchdog==4.0.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.4
wrapt==1.12.1
xformer==1.0.1
xformers==0.0.28.post1
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
